260: Fix OAuth FAPI issues (#70)
The aim here is to get a working flow that is FAPI compliant. We can't
do that as yet, but this moves us to a place were at least we can get
flows working using postman. We are now in a position where the URI on
which the OAuth2 well_know response is obtained matches the issuer url
within the response. To do this we have configured the OAuth2Provider to
accept obdemo.<namespace>.forgerock.financial as an Acceptable Audience
Value. This is all good.

The current issue is that the authorize endpoint must be accessed on the
iam.<namespace> domain, but the FAPI conformance suite uses the issuer
value from the well_known namespace as the audience in the request jwt.
This is now the obdemo.<namespace> domain. :-( We don't have an easy way
to fix this.

Issue: https://github.com/securebankingaccesstoolkit/securebankingaccesstoolkit/issues/260