441: Removing environment.namespace value (#95)
Currently, the cronjob-secret is being created in the dev namespace (default value) for the nightly environment (because we are not overriding the value).

I think that we can simply remove the environment.namespace value and allow Argo CD to default the namespace to that of the application. i.e. All objects for the nightly application will be placed in the nightly namespace.

Issue: https://github.com/SecureBankingAccessToolkit/SecureBankingAccessToolkit/issues/441